third_party/blink/web_tests/external/wpt/editing/run/forwarddelete_7001-last-expected.txt

This is a testharness.js-based test.
Found 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font size=3>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"3\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font size=\\"3\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font size=4>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"4\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font size=\\"4\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font size=4>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"4\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font size=\\"4\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font color=blue>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font color=blue>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font size=5>foo[]</font><p><font color=blue>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"5\\">foo</font><font color=\\"blue\\">bar</font></p>" but got "<p><font size=\\"5\\">foo</font><span style=\\"color:rgb(0, 0, 255)\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font size=5>foo[]</font><p><font color=blue>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"5\\">foo</font><font color=\\"blue\\">bar</font></p>" but got "<p><font size=\\"5\\">foo</font><span style=\\"color:rgb(0, 0, 255)\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font face=monospace>foo[]</font><p><font face=sans-serif>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font face=\\"monospace\\">foo</font><font face=\\"sans-serif\\">bar</font></p>" but got "<p><font face=\\"monospace\\">foo</font><span style=\\"font-family:sans-serif\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font face=monospace>foo[]</font><p><font face=sans-serif>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font face=\\"monospace\\">foo</font><font face=\\"sans-serif\\">bar</font></p>" but got "<p><font face=\\"monospace\\">foo</font><span style=\\"font-family:sans-serif\\">bar</span></p>"
[FAIL] [["forwarddelete",""]] "<div><span style=\\"display:list-item\\">abc[]</span><span style=\\"display:list-item\\">def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><span style=\\"display:list-item\\">abcdef</span></div>" but got "<span style=\\"display:list-item\\">abcdef</span>"
[FAIL] [["forwarddelete",""]] "<div>{abc</div><div>def</div>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div>" but got "<br>"
Harness: the test ran to completion.
